loginauth_onetomany

+++ app/Http/Controllers/TasksController.php
@@ -15,10 +15,17 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        return view('tasks.index', [
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc' )->paginate(10);
+            
+            return view('tasks.index', [
             'tasks' =>$tasks,
             ]);
+        }
+        else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -50,8 +57,10 @@ class TasksController extends Controller
         
         $task = new Task;
         $task->status = $request->status;
+        $task->user_id = \Auth::user()->id;
         $task->content = $request->content;
         $task->save();
+
 
         return redirect('/');
     }
@@ -64,12 +73,21 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::find($id);
-
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
-    }
+        if (\Auth::check()) {
+           $user = \Auth::user();
+           $task = Task::find($id);
+            if ($task->user_id == $user->id) {
+                return view('tasks.show', [
+                   'task' => $task
+                ]);
+            } else {
+              return redirect('/');
+            }
+        }else {
+            return view('welcome');
+        }
+    }        
+        
 
     /**
      * Show the form for editing the specified resource.
@@ -117,8 +135,11 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        if (\Auth::user()->id === $task->user_id) {
+              $task = Task::find($id);
+              $task->delete();
+        }
+
 
         return redirect('/');
     }
